pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
[project]
name = "dae"
version = "0.0.1"
dependencies = [
    "numpy",
    'importlib-metadata; python_version<"3.8"',
    "torch",
    "torchvision",
    "pandas",
    "hydra-core",
    "matplotlib",
    "seaborn"
]

